Refactor: preview_post.dart
