+++ lib/pages/main_page.dart
@@ -1,9 +1,12 @@
+// ignore_for_file: must_be_immutable
+
 import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:new_ara_app/constants/colors_info.dart';
 import 'package:new_ara_app/providers/user_provider.dart';
 import 'package:new_ara_app/widgetclasses/loading_indicator.dart';
+import 'package:new_ara_app/widgetclasses/preview_post.dart';
 import 'package:provider/provider.dart';
 
 class MainPage extends StatefulWidget {
@@ -427,22 +430,6 @@ class PopularBoard extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    DateTime now = DateTime.now();
-
-    DateTime date = DateTime.parse(json["created_at"]).toLocal();
-    var difference = now.difference(date);
-    String time = "미정";
-    if (difference.inMinutes < 1) {
-      time = "${difference.inSeconds}초 전";
-    } else if (difference.inHours < 1) {
-      time = '${difference.inMinutes} 분전';
-    } else if (difference.inDays < 1) {
-      time =
-          '${DateFormat('HH').format(date)}:${DateFormat('mm').format(date)}';
-    } else {
-      time =
-          '${DateFormat('MM').format(date)}/${DateFormat('dd').format(date)}';
-    }
 
     return Container(
       // height: 100,
@@ -470,128 +457,16 @@ class PopularBoard extends StatelessWidget {
             width: 15,
           ),
           Expanded(
-            child: Column(
-              mainAxisAlignment: MainAxisAlignment.center,
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Row(
-                  children: [
-                    Expanded(
-                      child: Text(
-                        json["title"],
-                        style: const TextStyle(
-                          fontWeight: FontWeight.bold,
-                          fontSize: 16,
-                        ),
-                        maxLines: 1,
-                        overflow: TextOverflow.ellipsis,
-                      ),
-                    ),
-                    SizedBox(
-                      height: 14.22,
-                      child: json["attachment_type"] == "IMAGE"
-                          ? SvgPicture.asset(
-                              'assets/icons/has-picture.svg',
-                              fit: BoxFit.fitHeight,
-                            )
-                          : json["attachment_type"] == "NON_IMAGE"
-                              ? SvgPicture.asset(
-                                  'assets/icons/has-picture.svg',
-                                  fit: BoxFit.fitHeight,
-                                )
-                              : Container(),
-                    )
-                  ],
-                  //attachment_type
-                ),
-                SizedBox(
-                  height: 17.7,
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                    children: [
-                      Row(
-                        children: [
-                          Text(
-                            json["created_by"]["profile"]["nickname"],
-                            maxLines: 1,
-                            overflow: TextOverflow.ellipsis,
-                            style: const TextStyle(
-                              fontSize: 12,
-                              color: Color(0xFFB1B1B1),
-                            ),
-                          ),
-                          const SizedBox(
-                            width: 8,
-                          ),
-                          Text(
-                            time,
-                            maxLines: 1,
-                            style: const TextStyle(
-                                fontSize: 12, color: Color(0xFFB1B1B1)),
-                          ),
-                        ],
-                      ),
-                      Row(
-                        children: [
-                          SvgPicture.asset(
-                            'assets/icons/like.svg',
-                            fit: BoxFit.cover,
-                          ),
-                          Text(
-                            json["positive_vote_count"].toString(),
-                            style: const TextStyle(
-                              fontWeight: FontWeight.w500,
-                              fontSize: 12,
-                              color: Color(0xFFED3A3A),
-                            ),
-                            textAlign: TextAlign.center,
-                          ),
-                          const SizedBox(
-                            width: 7,
-                          ),
-                          SvgPicture.asset(
-                            'assets/icons/dislike.svg',
-                            fit: BoxFit.cover,
-                          ),
-                          Text(
-                            json["negative_vote_count"].toString(),
-                            style: const TextStyle(
-                                fontWeight: FontWeight.w500,
-                                fontSize: 12,
-                                color: Color(0xFF538DD1)),
-                            textAlign: TextAlign.center,
-                          ),
-                          const SizedBox(
-                            width: 7,
-                          ),
-                          SvgPicture.asset(
-                            'assets/icons/comment.svg',
-                            fit: BoxFit.cover,
-                          ),
-                          const SizedBox(
-                            width: 3,
-                          ),
-                          Text(
-                            json["comment_count"].toString(),
-                            style: const TextStyle(
-                              fontWeight: FontWeight.w500,
-                              fontSize: 12,
-                              color: Color(0xFF636363),
-                            ),
-                          ),
-                        ],
-                      )
-                    ],
-                  ),
-                ),
-              ],
-            ),
+            child: PreviewPost(json:json)
           ),
         ],
       ),
     );
   }
 }
+
+
+
 
 class MainPageTextButton extends StatelessWidget {
   final String buttonTitle;
